Update Velope.java
Added functionality to choose which player will be sent to chosen server. 

So if you create a command /survival and if you have the permission "velope.teleportplayers" you will be able to send a player to chosen group of servers by adding as an argument players nickname. It works also in the console.